# apps/backend/swagger.json
{"openapi":"3.0.0","paths":{"/auth/discord":{"get":{"operationId":"AuthController_discordAuth","parameters":[],"responses":{"302":{"description":"Discordの認証ページにリダイレクト"}},"summary":"Discordログイン開始","tags":["Auth"]}},"/auth/discord/callback":{"get":{"operationId":"AuthController_discordAuthCallback","parameters":[{"name":"error","required":true,"in":"query","schema":{"type":"string"}},{"name":"redirect","required":true,"in":"query","schema":{"type":"string"}},{"name":"code","required":true,"in":"query","description":"Discord認証コード","schema":{"type":"string"}}],"responses":{"200":{"description":"Discord認証成功","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}}}}}}},"summary":"Discord認証コールバック","tags":["Auth"]}},"/auth/verify":{"get":{"operationId":"AuthController_verifyToken","parameters":[],"responses":{"200":{"description":"トークンが有効","content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string","description":"Discord User ID"}}}}}},"401":{"description":"無効なトークン"}},"security":[{"token":[]}],"summary":"JWTトークンの検証","tags":["Auth"]}},"/auth/servers":{"get":{"operationId":"AuthController_getDiscordServers","parameters":[],"responses":{"200":{"description":"Discordサーバー取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserServersResponse"}}}}},"summary":"Discordサーバー取得","tags":["Auth"]}},"/user":{"get":{"operationId":"UserController_findAll","parameters":[{"name":"name","required":false,"in":"query","description":"ユーザー名","schema":{"type":"string"}},{"name":"take","required":false,"in":"query","description":"取得件数","schema":{"minimum":1,"type":"number"}},{"name":"skip","required":false,"in":"query","description":"取得開始位置","schema":{"minimum":0,"type":"number"}}],"responses":{"200":{"description":"検索条件に当てはまるユーザの取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindAllUser"}}}}},"summary":"検索条件に当てはまるユーザの全取得","tags":["User"]}},"/user/login":{"get":{"operationId":"UserController_login","parameters":[],"responses":{"200":{"description":"最終ログイン時間の更新と自身のデータ取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized"}},"summary":"認証済みでのログイン","tags":["User"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"指定したユーザの取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found"}},"summary":"指定したユーザの取得","tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"ユーザの更新成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"summary":"ユーザの更新","tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[],"responses":{"200":{"description":"ユーザの削除成功"}},"summary":"ユーザの削除","tags":["User"]}},"/servers/join":{"post":{"operationId":"ServersController_join","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinServerDto"}}}},"responses":{"200":{"description":"サーバーに参加しました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerWithServerUser"}}}}},"summary":"サーバーに参加する","tags":["Servers"]}},"/servers/me/server-user":{"get":{"operationId":"ServersController_findMeServerUser","parameters":[{"name":"serverId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"サーバーに参加しているかどうか","content":{"application/json":{"schema":{"type":"boolean"}}}}},"summary":"自身がサーバーに参加しているかどうかを取得する","tags":["Servers"]}},"/servers":{"get":{"operationId":"ServersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Servers"]}},"/servers/{id}":{"get":{"operationId":"ServersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Servers"]},"patch":{"operationId":"ServersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServerDto"}}}},"responses":{"200":{"description":""}},"tags":["Servers"]},"delete":{"operationId":"ServersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Servers"]}},"/calendars":{"post":{"operationId":"CalendarsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCalendarDto"}}}},"responses":{"201":{"description":"カレンダー作成成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalendarWithRelations"}}}}},"tags":["Calendars"]},"get":{"operationId":"CalendarsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Calendars"]}},"/calendars/{id}":{"get":{"operationId":"CalendarsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"カレンダー取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalendarWithRelations"}}}}},"tags":["Calendars"]},"patch":{"operationId":"CalendarsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCalendarDto"}}}},"responses":{"200":{"description":""}},"tags":["Calendars"]},"delete":{"operationId":"CalendarsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Calendars"]}},"/schedules/public":{"post":{"operationId":"SchedulesController_createPublicSchedule","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePublicScheduleDto"}}}},"responses":{"200":{"description":"公開スケジュール作成成功"}},"summary":"公開スケジュール作成","tags":["Schedules"]}},"/schedules/personal":{"post":{"operationId":"SchedulesController_createPersonalSchedule","parameters":[],"requestBody":{"required":true,"description":"個人スケジュール作成リクエスト","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreatePersonalScheduleDto"}}}}},"responses":{"201":{"description":"個人スケジュール作成成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonalSchedule"}}}}},"summary":"個人スケジュール作成","tags":["Schedules"]}},"/schedules":{"get":{"operationId":"SchedulesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Schedules"]}},"/schedules/{id}":{"get":{"operationId":"SchedulesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Schedules"]},"patch":{"operationId":"SchedulesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateScheduleDto"}}}},"responses":{"200":{"description":""}},"tags":["Schedules"]},"delete":{"operationId":"SchedulesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Schedules"]}}},"info":{"title":"GFC Scheduler API","description":"GFC Scheduler API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"Calendar":{"type":"object","properties":{"id":{"type":"string","description":"カレンダーID"},"serverId":{"type":"string","description":"サーバーID"},"name":{"type":"string","description":"カレンダー名"},"createdAt":{"format":"date-time","type":"string","description":"作成日"},"updatedAt":{"format":"date-time","type":"string","description":"更新日"}},"required":["id","serverId","name","createdAt","updatedAt"]},"ServerUser":{"type":"object","properties":{"userId":{"type":"string","description":"サーバーに参加しているユーザーID"},"serverId":{"type":"string","description":"サーバーID"},"createdAt":{"format":"date-time","type":"string","description":"参加日時"},"updatedAt":{"format":"date-time","type":"string","description":"更新日時"}},"required":["userId","serverId","createdAt","updatedAt"]},"ServerWithRelations":{"type":"object","properties":{"id":{"type":"string","description":"サーバーID"},"name":{"type":"string","description":"サーバー名"},"icon":{"type":"string","description":"サーバーアイコンURL","nullable":true},"calendars":{"description":"カレンダー","type":"array","items":{"$ref":"#/components/schemas/Calendar"}},"serverUsers":{"description":"サーバーに参加しているユーザー","type":"array","items":{"$ref":"#/components/schemas/ServerUser"}}},"required":["id","name","icon","calendars","serverUsers"]},"GetUserServersResponse":{"type":"object","properties":{"data":{"description":"サーバー","type":"array","items":{"$ref":"#/components/schemas/ServerWithRelations"}},"calendarCount":{"type":"number","description":"カレンダー数"}},"required":["data","calendarCount"]},"User":{"type":"object","properties":{"id":{"type":"string","description":"DiscordユーザID"},"name":{"type":"string","description":"ユーザー名"},"avatar":{"type":"string","description":"アバター画像URL","nullable":true},"createdAt":{"format":"date-time","type":"string","description":"作成日時"},"updatedAt":{"format":"date-time","type":"string","description":"更新日時"},"lastLoggedInAt":{"format":"date-time","type":"string","description":"最終ログイン日時"}},"required":["id","name","avatar","createdAt","updatedAt","lastLoggedInAt"]},"FindAllUser":{"type":"object","properties":{"users":{"description":"ユーザー一覧","type":"array","items":{"$ref":"#/components/schemas/User"}},"total":{"type":"number","description":"総件数"}},"required":["users","total"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string","description":"ユーザ名"}}},"JoinServerDto":{"type":"object","properties":{"serverId":{"type":"string","description":"参加するサーバーのID"},"serverName":{"type":"string","description":"参加するサーバーの名前"},"serverIcon":{"type":"string","description":"参加するサーバーのアイコン"}},"required":["serverId","serverName","serverIcon"]},"ServerWithServerUser":{"type":"object","properties":{"id":{"type":"string","description":"サーバーID"},"name":{"type":"string","description":"サーバー名"},"icon":{"type":"string","description":"サーバーアイコンURL","nullable":true},"serverUser":{"description":"サーバーに参加しているユーザー","allOf":[{"$ref":"#/components/schemas/ServerUser"}]}},"required":["id","name","icon","serverUser"]},"UpdateServerDto":{"type":"object","properties":{"name":{"type":"string","description":"サーバー名","example":"My Server"},"icon":{"type":"string","description":"サーバーアイコン","example":"ABCDEF12345"},"serverId":{"type":"string","description":"サーバーID","example":"1234567890"}}},"CreateCalendarDto":{"type":"object","properties":{"name":{"type":"string","description":"カレンダー名"},"serverId":{"type":"string","description":"サーバーID"},"serverName":{"type":"string","description":"サーバー名"},"icon":{"type":"string","description":"サーバーアイコンURL","nullable":true}},"required":["name","serverId","serverName","icon"]},"Server":{"type":"object","properties":{"id":{"type":"string","description":"サーバーID"},"name":{"type":"string","description":"サーバー名"},"icon":{"type":"string","description":"サーバーアイコンURL","nullable":true}},"required":["id","name","icon"]},"ScheduleUserInfo":{"type":"object","properties":{"id":{"type":"string","description":"ユーザーID"},"name":{"type":"string","description":"ユーザー名"},"avatar":{"type":"string","description":"アバターURL"}},"required":["id","name","avatar"]},"Participant":{"type":"object","properties":{"userId":{"type":"string","description":"参加者ID"},"name":{"type":"string","description":"参加者名"},"reaction":{"type":"string","description":"参加者反応","enum":["OK","UNDECIDED","NG"]},"createdAt":{"format":"date-time","type":"string","description":"参加者反応日"},"updatedAt":{"format":"date-time","type":"string","description":"参加者反応更新日"}},"required":["userId","name","reaction","createdAt","updatedAt"]},"PublicScheduleWithRelations":{"type":"object","properties":{"id":{"type":"string","description":"ID"},"date":{"type":"string","description":"日付"},"title":{"type":"string","description":"タイトル"},"description":{"type":"string","description":"説明"},"createdBy":{"description":"作成者情報","allOf":[{"$ref":"#/components/schemas/ScheduleUserInfo"}]},"updatedBy":{"description":"更新者情報","allOf":[{"$ref":"#/components/schemas/ScheduleUserInfo"}]},"createdById":{"type":"string","description":"作成者ID"},"updatedById":{"type":"string","description":"更新者ID"},"createdAt":{"format":"date-time","type":"string","description":"作成日"},"updatedAt":{"format":"date-time","type":"string","description":"更新日"},"recruitCount":{"type":"number","description":"募集人数"},"isRecruiting":{"type":"boolean","description":"募集状況"},"isPersonal":{"type":"boolean","description":"個人予定フラグ","default":false},"participants":{"description":"参加者","type":"array","items":{"$ref":"#/components/schemas/Participant"}}},"required":["id","date","title","description","createdBy","updatedBy","createdById","updatedById","createdAt","updatedAt","recruitCount","isRecruiting","isPersonal","participants"]},"PersonalScheduleWithRelations":{"type":"object","properties":{"id":{"type":"string","description":"ID"},"date":{"type":"string","description":"日付"},"title":{"type":"string","description":"タイトル"},"description":{"type":"string","description":"説明"},"createdBy":{"description":"作成者情報","allOf":[{"$ref":"#/components/schemas/ScheduleUserInfo"}]},"updatedBy":{"description":"更新者情報","allOf":[{"$ref":"#/components/schemas/ScheduleUserInfo"}]},"createdById":{"type":"string","description":"作成者ID"},"updatedById":{"type":"string","description":"更新者ID"},"createdAt":{"format":"date-time","type":"string","description":"作成日"},"updatedAt":{"format":"date-time","type":"string","description":"更新日"},"isPrivate":{"type":"boolean","description":"非公開フラグ"},"isFree":{"type":"boolean","description":"空き予定フラグ"},"isPersonal":{"type":"boolean","description":"個人予定フラグ","default":true},"user":{"description":"作成したユーザ","allOf":[{"$ref":"#/components/schemas/User"}]}},"required":["id","date","title","description","createdBy","updatedBy","createdById","updatedById","createdAt","updatedAt","isPrivate","isFree","isPersonal","user"]},"CalendarWithRelations":{"type":"object","properties":{"id":{"type":"string","description":"カレンダーID"},"serverId":{"type":"string","description":"サーバーID"},"name":{"type":"string","description":"カレンダー名"},"createdAt":{"format":"date-time","type":"string","description":"作成日"},"updatedAt":{"format":"date-time","type":"string","description":"更新日"},"server":{"description":"サーバー","allOf":[{"$ref":"#/components/schemas/Server"}]},"publicSchedules":{"description":"公開予定","type":"array","items":{"$ref":"#/components/schemas/PublicScheduleWithRelations"}},"personalSchedules":{"description":"個人予定","type":"array","items":{"$ref":"#/components/schemas/PersonalScheduleWithRelations"}}},"required":["id","serverId","name","createdAt","updatedAt","server","publicSchedules","personalSchedules"]},"UpdateCalendarDto":{"type":"object","properties":{"name":{"type":"string","description":"カレンダー名"},"serverId":{"type":"string","description":"サーバーID"},"serverName":{"type":"string","description":"サーバー名"},"icon":{"type":"string","description":"サーバーアイコンURL","nullable":true}}},"CreatePublicScheduleDto":{"type":"object","properties":{"date":{"format":"date-time","type":"string","description":"日付"},"title":{"type":"string","description":"タイトル"},"description":{"type":"string","description":"説明"},"recruitCount":{"type":"number","description":"募集人数"}},"required":["date","title","description","recruitCount"]},"CreatePersonalScheduleDto":{"type":"object","properties":{"date":{"type":"string","description":"日付"},"title":{"type":"string","description":"タイトル"},"description":{"type":"string","description":"説明"},"isPrivate":{"type":"boolean","description":"非公開フラグ"},"isFree":{"type":"boolean","description":"無料フラグ"}},"required":["date","title","description","isPrivate","isFree"]},"PersonalSchedule":{"type":"object","properties":{"id":{"type":"string","description":"ID"},"date":{"type":"string","description":"日付"},"title":{"type":"string","description":"タイトル"},"description":{"type":"string","description":"説明"},"createdBy":{"description":"作成者情報","allOf":[{"$ref":"#/components/schemas/ScheduleUserInfo"}]},"updatedBy":{"description":"更新者情報","allOf":[{"$ref":"#/components/schemas/ScheduleUserInfo"}]},"createdById":{"type":"string","description":"作成者ID"},"updatedById":{"type":"string","description":"更新者ID"},"createdAt":{"format":"date-time","type":"string","description":"作成日"},"updatedAt":{"format":"date-time","type":"string","description":"更新日"},"isPrivate":{"type":"boolean","description":"非公開フラグ"},"isFree":{"type":"boolean","description":"空き予定フラグ"},"isPersonal":{"type":"boolean","description":"個人予定フラグ","default":true},"user":{"description":"作成したユーザ","allOf":[{"$ref":"#/components/schemas/User"}]}},"required":["id","date","title","description","createdBy","updatedBy","createdById","updatedById","createdAt","updatedAt","isPrivate","isFree","isPersonal","user"]},"UpdateScheduleDto":{"type":"object","properties":{}}}}}